refactor: remove provider registry adapter module
